+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Changed
+- Connecting to internet servers without GameSpy account is now possible.
+
 ## [2.5] - 2020-12-22
 ### Fixed
 - Unlocked memory above 2 GB in 32-bit version. Thanks to Vladislav.